refactor: 유저 Controller에 있는 다른 도메인 API 분리 (#128)
