+++ src/main/java/com/jisungin/api/user/UserController.java
@@ -2,18 +2,10 @@ package com.jisungin.api.user;
 
 import com.jisungin.api.ApiResponse;
 import com.jisungin.api.support.Auth;
-import com.jisungin.api.user.request.ReviewContentGetAllRequest;
-import com.jisungin.api.user.request.UserRatingGetAllRequest;
-import com.jisungin.api.user.request.UserReadingStatusGetAllRequest;
-import com.jisungin.application.PageResponse;
-import com.jisungin.application.rating.response.RatingGetResponse;
-import com.jisungin.application.review.response.ReviewContentGetAllResponse;
 import com.jisungin.application.user.UserService;
 import com.jisungin.application.user.response.UserInfoResponse;
-import com.jisungin.application.library.response.UserReadingStatusResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -24,43 +16,9 @@ public class UserController {
 
     private final UserService userService;
 
-    @GetMapping("/ratings")
-    public ApiResponse<PageResponse<RatingGetResponse>> getUserRatings(
-            @ModelAttribute UserRatingGetAllRequest request,
-            @Auth Long userId
-    ) {
-        PageResponse<RatingGetResponse> response = userService.getUserRatings(
-                userId, request.toService());
-
-        return ApiResponse.ok(response);
-    }
-
-    @GetMapping("/reviews")
-    public ApiResponse<ReviewContentGetAllResponse> getReviewContents(
-            @ModelAttribute ReviewContentGetAllRequest request,
-            @Auth Long userId
-    ) {
-        ReviewContentGetAllResponse response = userService.getReviewContents(
-                userId, request.toService());
-
-        return ApiResponse.ok(response);
-    }
-
-    @GetMapping("/statuses")
-    public ApiResponse<PageResponse<UserReadingStatusResponse>> getReadingStatuses(
-            @ModelAttribute UserReadingStatusGetAllRequest request,
-            @Auth Long userId
-    ) {
-        PageResponse<UserReadingStatusResponse> response = userService
-                .getUserReadingStatuses(userId, request.toService());
-
-        return ApiResponse.ok(response);
-    }
-
     @GetMapping("/me")
     public ApiResponse<UserInfoResponse> getUserInfo(@Auth Long userId) {
         UserInfoResponse userInfo = userService.getUserInfo(userId);
         return ApiResponse.ok(userInfo);
     }
-
 }
