+++ src/main/java/com/jisungin/application/review/ReviewService.java
@@ -1,18 +1,18 @@
 package com.jisungin.application.review;
 
-import static com.jisungin.exception.ErrorCode.BOOK_NOT_FOUND;
-import static com.jisungin.exception.ErrorCode.REVIEW_NOT_FOUND;
-import static com.jisungin.exception.ErrorCode.UNAUTHORIZED_REQUEST;
-import static com.jisungin.exception.ErrorCode.USER_NOT_FOUND;
-
 import com.jisungin.application.OffsetLimit;
+import com.jisungin.application.PageResponse;
 import com.jisungin.application.SliceResponse;
+import com.jisungin.application.review.request.ReviewContentGetAllServiceRequest;
 import com.jisungin.application.review.request.ReviewCreateServiceRequest;
+import com.jisungin.application.review.response.ReviewContentGetAllResponse;
+import com.jisungin.application.review.response.ReviewContentResponse;
 import com.jisungin.application.review.response.ReviewWithRatingResponse;
 import com.jisungin.domain.book.Book;
 import com.jisungin.domain.book.repository.BookRepository;
 import com.jisungin.domain.review.Review;
 import com.jisungin.domain.review.repository.ReviewRepository;
+import com.jisungin.domain.reviewlike.repository.ReviewLikeRepository;
 import com.jisungin.domain.user.User;
 import com.jisungin.domain.user.repository.UserRepository;
 import com.jisungin.exception.BusinessException;
@@ -20,12 +20,17 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
+import static com.jisungin.exception.ErrorCode.*;
+
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 @Service
 public class ReviewService {
 
     private final ReviewRepository reviewRepository;
+    private final ReviewLikeRepository reviewLikeRepository;
     private final UserRepository userRepository;
     private final BookRepository bookRepository;
 
@@ -70,4 +75,19 @@ public class ReviewService {
         reviewRepository.delete(deleteReview);
     }
 
+    public ReviewContentGetAllResponse getReviewContents(Long userId, ReviewContentGetAllServiceRequest request) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new BusinessException(USER_NOT_FOUND));
+
+        PageResponse<ReviewContentResponse> reviewContents = reviewRepository.findAllReviewContentOrderBy(
+                user.getId(), request.getOrderType(), request.getSize(), request.getOffset());
+
+        List<Long> reviewIds = reviewContents.getQueryResponse().stream()
+                .map(ReviewContentResponse::getReviewId)
+                .toList();
+
+        List<Long> likeReviewIds = reviewLikeRepository.findLikeReviewByReviewId(userId, reviewIds);
+
+        return ReviewContentGetAllResponse.of(reviewContents, likeReviewIds);
+    }
 }
