+++ src/test/java/com/jisungin/ServiceTestSupport.java
@@ -13,5 +13,4 @@ public abstract class ServiceTestSupport {
 
     @MockBean
     protected ClientRegistrationRepository clientRegistrationRepository;
-
 }
