+++ src/test/java/com/jisungin/application/review/ReviewServiceTest.java
@@ -1,15 +1,16 @@
 package com.jisungin.application.review;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
 import com.jisungin.ServiceTestSupport;
 import com.jisungin.application.OffsetLimit;
 import com.jisungin.application.SliceResponse;
 import com.jisungin.application.review.request.ReviewCreateServiceRequest;
+import com.jisungin.application.review.response.ReviewContentGetAllResponse;
 import com.jisungin.application.review.response.ReviewWithRatingResponse;
+import com.jisungin.application.review.request.ReviewContentGetAllServiceRequest;
 import com.jisungin.domain.book.Book;
 import com.jisungin.domain.book.repository.BookRepository;
+import com.jisungin.domain.rating.Rating;
+import com.jisungin.domain.rating.repository.RatingRepository;
 import com.jisungin.domain.review.Review;
 import com.jisungin.domain.review.repository.ReviewRepository;
 import com.jisungin.domain.reviewlike.ReviewLike;
@@ -19,13 +20,20 @@ import com.jisungin.domain.user.OauthType;
 import com.jisungin.domain.user.User;
 import com.jisungin.domain.user.repository.UserRepository;
 import com.jisungin.exception.BusinessException;
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.stream.IntStream;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+import static com.jisungin.domain.review.RatingOrderType.RATING_AVG_ASC;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.groups.Tuple.tuple;
 
 class ReviewServiceTest extends ServiceTestSupport {
 
@@ -34,6 +42,9 @@ class ReviewServiceTest extends ServiceTestSupport {
 
     @Autowired
     private ReviewLikeRepository reviewLikeRepository;
+
+    @Autowired
+    private RatingRepository ratingRepository;
 
     @Autowired
     private UserRepository userRepository;
@@ -46,6 +57,7 @@ class ReviewServiceTest extends ServiceTestSupport {
 
     @AfterEach
     void tearDown() {
+        ratingRepository.deleteAllInBatch();
         reviewLikeRepository.deleteAllInBatch();
         reviewRepository.deleteAllInBatch();
         bookRepository.deleteAllInBatch();
@@ -211,6 +223,66 @@ class ReviewServiceTest extends ServiceTestSupport {
                 .hasMessage("권한이 없는 사용자입니다.");
     }
 
+    @DisplayName("사용자의 모든 리뷰 내용을 별점 오름차 순으로 가져온다.")
+    @Test
+    void getReviewContents() {
+        //given
+        User user1 = userRepository.save(createUser("1"));
+        User user2 = userRepository.save(createUser("2"));
+        List<Book> books = bookRepository.saveAll(createBooks());
+        List<Review> reviews = reviewRepository.saveAll(createReviews(user1, books));
+        List<Rating> ratings1 = ratingRepository.saveAll(createRatings(user1, books));
+        List<Rating> ratings2 = ratingRepository.saveAll(createRatings(user2, books));
+        List<ReviewLike> reviewLikesWithUser1 = reviewLikeRepository.saveAll(createReviewLikes(user1, reviews));
+        List<ReviewLike> reviewLikesWithUser2 = reviewLikeRepository.saveAll(createReviewLikes(user2, reviews));
+
+        // 1번째부터 4번째 데이터를 요청, 별점 1.0인 리뷰 내용이 4개 나와야 함
+        ReviewContentGetAllServiceRequest request = ReviewContentGetAllServiceRequest.builder()
+                .page(1)
+                .size(4)
+                .orderType(RATING_AVG_ASC)
+                .build();
+
+        //when
+        ReviewContentGetAllResponse result = reviewService.getReviewContents(user1.getId(), request);
+
+        //then
+        assertThat(result.getReviewContents().getTotalCount()).isEqualTo(20);
+        assertThat(result.getReviewContents().getQueryResponse()).hasSize(4)
+                .extracting(
+                        "userImage", "userName", "rating", "content", "isbn", "title", "bookImage")
+                .containsExactly(
+                        tuple("image", "김도형", 1.0, "리뷰 내용1", "1", "제목1", "bookImage"),
+                        tuple("image", "김도형", 1.0, "리뷰 내용11", "11", "제목11", "bookImage"),
+                        tuple("image", "김도형", 1.0, "리뷰 내용16", "16", "제목16", "bookImage"),
+                        tuple("image", "김도형", 1.0, "리뷰 내용6", "6", "제목6", "bookImage")
+                );
+        assertThat(result.getUserLikes()).hasSize(4);
+    }
+
+    private static List<Review> createReviews(User user, List<Book> books) {
+        return IntStream.range(0, 20)
+                .mapToObj(i -> {
+                    double rating = i % 5 + 1.0; // 1.0, 2.0, 3.0, 4.0, 5.0이 순환되도록 설정
+                    return createReview(user, books.get(i), rating); // Review 객체 생성
+                })
+                .collect(Collectors.toList());
+    }
+
+    private static Review createReview(User user, Book book, Double rating) {
+        return Review.builder()
+                .user(user)
+                .book(book)
+                .content("리뷰 내용" + book.getIsbn())
+                .build();
+    }
+
+    private static List<ReviewLike> createReviewLikes(User user, List<Review> reviews) {
+        return reviews.stream()
+                .map(review -> createReviewLike(user, review))
+                .toList();
+    }
+
     private static ReviewLike createReviewLike(User user, Review review) {
         return ReviewLike.builder()
                 .user(user)
@@ -270,4 +342,39 @@ class ReviewServiceTest extends ServiceTestSupport {
                 .build();
     }
 
+    private static List<Book> createBooks() {
+        return IntStream.rangeClosed(1, 20)
+                .mapToObj(i -> createBook(
+                        "제목" + String.valueOf(i), "내용" + String.valueOf(i), String.valueOf(i)))
+                .collect(Collectors.toList());
+    }
+
+    private static Book createBook(String title, String content, String isbn) {
+        return Book.builder()
+                .title(title)
+                .content(content)
+                .authors("김도형")
+                .isbn(isbn)
+                .publisher("지성인")
+                .dateTime(LocalDateTime.of(2024, 1, 1, 0, 0))
+                .imageUrl("bookImage")
+                .build();
+    }
+
+    private static List<Rating> createRatings(User user, List<Book> books) {
+        return IntStream.range(0, 20)
+                .mapToObj(i -> {
+                    double rating = i % 5 + 1.0;
+                    return createRating(user, books.get(i), rating);
+                })
+                .collect(Collectors.toList());
+    }
+
+    private static Rating createRating(User user, Book book, Double rating) {
+        return Rating.builder()
+                .user(user)
+                .book(book)
+                .rating(rating)
+                .build();
+    }
 }
